requirements.txt
# UltraFaceswap - Face Swap Pipeline Dependencies
# Python 3.10+ required

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Face detection and swap (InsightFace)
insightface>=0.7.3
onnxruntime>=1.16.0
# For GPU: pip install onnxruntime-gpu

# Face restoration (optional - install separately if needed)
# gfpgan>=1.3.8

# Video processing
tqdm>=4.66.0

# Video processing (FFmpeg must be installed separately)
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# API (Phase 2)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Queue (Phase 2)
celery>=5.3.0
redis>=5.0.0
